+++ appserver/neo4japp/schemas/projects.py
@@ -39,7 +39,7 @@ class ProjectMultiCollaboratorUpdateRequest(CamelCaseSchema):
     # We name this Multi- instead of Bulk- because it doesn't match the other
     # bulk semantics used in other schema (like the bulk file change request)
     update_or_create = fields.List(fields.Nested(ProjectCollaboratorUpdateSchema),
-                                   missing=lambda: [],
+                                   load_default=list,
                                    validate=marshmallow.validate.Length(max=5))
-    remove_user_hash_ids = fields.List(fields.String(), missing=lambda: [],
+    remove_user_hash_ids = fields.List(fields.String(), load_default=list,
                                        validate=marshmallow.validate.Length(max=5))